refactor: only one use query ai
Removes `useQuery` and renames `useSuspenseQuery` to `useQuery`

- mainteinance overhead to maintain two APIs that achieve the same thing was too high

- docs get simplified to one approach only
