+++ packages/lix/lix-react-utils/src/hooks/use-query.test.tsx
@@ -1,106 +1,117 @@
 import { test, expect } from "vitest";
-import { renderHook, act, waitFor } from "@testing-library/react";
-import React from "react";
-import { useQuery, useQueryTakeFirst } from "./use-query.js";
+import { renderHook, waitFor, act } from "@testing-library/react";
+import React, { Suspense } from "react";
+import {
+	useQuery,
+	useQueryTakeFirst,
+	useQueryTakeFirstOrThrow,
+} from "./use-query.js";
 import { LixProvider } from "../provider.js";
 import { openLix, type KeyValue, type State } from "@lix-js/sdk";
 
-async function waitForUpdate<T>(
-	result: { readonly current: T },
-	selector: (v: T) => unknown,
-) {
-	const start = selector(result.current);
-	await new Promise<void>((resolve) => {
-		const id = setInterval(() => {
-			if (selector(result.current) !== start) {
-				clearInterval(id);
-				resolve();
-			}
-		}, 1); // 1 ms polling – negligible in tests
-	});
+// React Error Boundaries require class components - no functional equivalent exists
+class MockErrorBoundary extends React.Component<
+	{ children: React.ReactNode; onError?: (error: Error) => void },
+	{ hasError: boolean; error?: Error }
+> {
+	override state = { hasError: false, error: undefined };
+
+	// @ts-expect-error - type error
+	static override getDerivedStateFromError(error: Error) {
+		return { hasError: true, error };
+	}
+
+	override componentDidCatch(error: Error) {
+		this.props.onError?.(error);
+	}
+
+	override render() {
+		return this.state.hasError ? (
+			<div>Error occurred</div>
+		) : (
+			this.props.children
+		);
+	}
 }
 
-test("useQuery throws error when used outside LixProvider", () => {
+test("useSuspenseQuery throws error when used outside LixProvider", () => {
 	// We need to catch the error since it's thrown during render
 	expect(() => {
 		renderHook(() =>
 			useQuery((lix) => lix.db.selectFrom("key_value").selectAll()),
 		);
-	}).toThrow("useQuery must be used inside <LixProvider>.");
+	}).toThrow("useSuspenseQuery must be used inside <LixProvider>.");
 });
 
-test("useQuery returns loading state initially", async () => {
+test("useSuspenseQuery returns array with data using new API", async () => {
 	const lix = await openLix({});
+
 	const wrapper = ({ children }: { children: React.ReactNode }) => (
-		<LixProvider lix={lix}>{children}</LixProvider>
+		<LixProvider lix={lix}>
+			<Suspense fallback={<div>Loading...</div>}>
+				<MockErrorBoundary>{children}</MockErrorBoundary>
+			</Suspense>
+		</LixProvider>
 	);
 
-	const { result } = renderHook(
-		() =>
-			useQuery((lix) =>
-				lix.db
-					.selectFrom("key_value")
-					.selectAll()
-					.where("key", "like", "test_%"),
-			),
-		{ wrapper },
-	);
+	let hookResult: { current: State<KeyValue>[] };
 
-	expect(result.current.loading).toBe(true);
-	expect(result.current.data).toBeUndefined();
-	expect(result.current.error).toBe(null);
+	await act(async () => {
+		const { result } = renderHook(
+			() => {
+				const data = useQuery((lix) =>
+					lix.db
+						.selectFrom("key_value")
+						.selectAll()
+						.where("key", "like", "test_%"),
+				);
+				return data;
+			},
+			{ wrapper },
+		);
+		hookResult = result;
+	});
+
+	// Wait for suspense to resolve and data to be available
+	await waitFor(() => {
+		expect(Array.isArray(hookResult.current)).toBe(true);
+		expect(hookResult.current).toEqual([]); // No test keys initially
+	});
 
 	await lix.close();
 });
 
-test("useQuery returns data after initial load", async () => {
+test("useSuspenseQuery updates when data changes", async () => {
 	const lix = await openLix({});
 	const wrapper = ({ children }: { children: React.ReactNode }) => (
-		<LixProvider lix={lix}>{children}</LixProvider>
+		<LixProvider lix={lix}>
+			<Suspense fallback={<div>Loading...</div>}>
+				<MockErrorBoundary>{children}</MockErrorBoundary>
+			</Suspense>
+		</LixProvider>
 	);
 
-	const { result } = renderHook(
-		() =>
-			useQuery((lix) =>
-				lix.db
-					.selectFrom("key_value")
-					.selectAll()
-					.where("key", "like", "test_%"),
-			),
-		{ wrapper },
-	);
+	let hookResult: { current: State<KeyValue>[] };
 
-	// Wait for loading to become false
-	await waitForUpdate(result, (r) => r.loading);
-
-	expect(result.current.loading).toBe(false);
-	expect(result.current.data).toEqual([]); // No test keys initially
-	expect(result.current.error).toBe(null);
-
-	await lix.close();
-});
-
-test("useQuery updates when data changes", async () => {
-	const lix = await openLix({});
-	const wrapper = ({ children }: { children: React.ReactNode }) => (
-		<LixProvider lix={lix}>{children}</LixProvider>
-	);
-
-	const { result } = renderHook(
-		() =>
-			useQuery((lix) =>
-				lix.db
-					.selectFrom("key_value")
-					.selectAll()
-					.where("key", "like", "react_test_%"),
-			),
-		{ wrapper },
-	);
+	await act(async () => {
+		const { result } = renderHook(
+			() => {
+				const data = useQuery((lix) =>
+					lix.db
+						.selectFrom("key_value")
+						.selectAll()
+						.where("key", "like", "react_test_%"),
+				);
+				return data;
+			},
+			{ wrapper },
+		);
+		hookResult = result;
+	});
 
 	// Wait for initial empty data
 	await waitFor(() => {
-		expect(result.current.loading).toBe(false);
-		expect(result.current.data).toEqual([]);
+		expect(hookResult.current).toEqual([]);
 	});
 
 	// Insert a test key-value pair
@@ -113,8 +124,8 @@ test("useQuery updates when data changes", async () => {
 
 	// Check updated data
 	await waitFor(() => {
-		expect(result.current.data).toHaveLength(1);
-		expect(result.current.data?.[0]).toMatchObject({
+		expect(hookResult.current).toHaveLength(1);
+		expect(hookResult.current[0]).toMatchObject({
 			key: "react_test_key",
 			value: "test_value",
 		});
@@ -123,95 +134,7 @@ test("useQuery updates when data changes", async () => {
 	await lix.close();
 });
 
-test("useQuery handles query errors", async () => {
-	const lix = await openLix({});
-	const wrapper = ({ children }: { children: React.ReactNode }) => (
-		<LixProvider lix={lix}>{children}</LixProvider>
-	);
-
-	const { result } = renderHook(
-		() =>
-			useQuery((lix) =>
-				lix.db.selectFrom("non_existent_table" as never).selectAll(),
-			),
-		{ wrapper },
-	);
-
-	await waitFor(() => {
-		expect(result.current.loading).toBe(false);
-	});
-
-	expect(result.current.error).toBeInstanceOf(Error);
-	expect(result.current.data).toBeUndefined();
-
-	await lix.close();
-});
-
-test("useQuery multiple subscriptions work independently", async () => {
-	const lix = await openLix({});
-	const wrapper = ({ children }: { children: React.ReactNode }) => (
-		<LixProvider lix={lix}>{children}</LixProvider>
-	);
-
-	const { result: result1 } = renderHook(
-		() =>
-			useQuery((lix) =>
-				lix.db
-					.selectFrom("key_value")
-					.selectAll()
-					.where("key", "like", "multi_1_%"),
-			),
-		{ wrapper },
-	);
-
-	const { result: result2 } = renderHook(
-		() =>
-			useQuery((lix) =>
-				lix.db
-					.selectFrom("key_value")
-					.selectAll()
-					.where("key", "like", "multi_2_%"),
-			),
-		{ wrapper },
-	);
-
-	// Wait for both to load
-	await waitFor(() => {
-		expect(result1.current.loading).toBe(false);
-		expect(result2.current.loading).toBe(false);
-	});
-
-	// Insert data for first query
-	await act(async () => {
-		await lix.db
-			.insertInto("key_value")
-			.values({ key: "multi_1_key", value: "value_1" })
-			.execute();
-	});
-
-	// Only first query should update
-	await waitFor(() => {
-		expect(result1.current.data).toHaveLength(1);
-	});
-	expect(result2.current.data).toHaveLength(0);
-
-	// Insert data for second query
-	await act(async () => {
-		await lix.db
-			.insertInto("key_value")
-			.values({ key: "multi_2_key", value: "value_2" })
-			.execute();
-	});
-
-	// Now second query should update
-	await waitFor(() => {
-		expect(result2.current.data).toHaveLength(1);
-	});
-
-	await lix.close();
-});
-
-test("useQueryTakeFirst returns first item from array", async () => {
+test("useSuspenseQueryTakeFirst returns array with single item or undefined", async () => {
 	const lix = await openLix({});
 
 	// Insert test data
@@ -224,222 +147,400 @@ test("useQueryTakeFirst returns first item from array", async () => {
 		.execute();
 
 	const wrapper = ({ children }: { children: React.ReactNode }) => (
-		<LixProvider lix={lix}>{children}</LixProvider>
+		<LixProvider lix={lix}>
+			<Suspense fallback={<div>Loading...</div>}>
+				<MockErrorBoundary>{children}</MockErrorBoundary>
+			</Suspense>
+		</LixProvider>
 	);
 
-	const { result } = renderHook(
-		() =>
-			useQueryTakeFirst((lix) =>
-				lix.db
-					.selectFrom("key_value")
-					.selectAll()
-					.where("key", "like", "first_test_%")
-					.orderBy("key", "asc"),
-			),
-		{ wrapper },
-	);
+	let hookResult: { current: State<KeyValue> | undefined };
 
-	await waitFor(() => {
-		expect(result.current.loading).toBe(false);
+	await act(async () => {
+		const { result } = renderHook(
+			() => {
+				const data = useQueryTakeFirst((lix) =>
+					lix.db
+						.selectFrom("key_value")
+						.selectAll()
+						.where("key", "like", "first_test_%")
+						.orderBy("key", "asc"),
+				);
+				return data;
+			},
+			{ wrapper },
+		);
+		hookResult = result;
 	});
 
-	expect(result.current.data).toMatchObject({
-		key: "first_test_1",
-		value: "first",
+	await waitFor(() => {
+		expect(hookResult.current).toMatchObject({
+			key: "first_test_1",
+			value: "first",
+		});
 	});
 
 	await lix.close();
 });
 
-test("useQueryTakeFirst returns undefined for empty results", async () => {
+test("useSuspenseQueryTakeFirst returns undefined for empty results", async () => {
 	const lix = await openLix({});
 	const wrapper = ({ children }: { children: React.ReactNode }) => (
-		<LixProvider lix={lix}>{children}</LixProvider>
+		<LixProvider lix={lix}>
+			<Suspense fallback={<div>Loading...</div>}>
+				<MockErrorBoundary>{children}</MockErrorBoundary>
+			</Suspense>
+		</LixProvider>
 	);
 
-	const { result } = renderHook(
-		() =>
-			useQueryTakeFirst((lix) =>
-				lix.db
-					.selectFrom("key_value")
-					.selectAll()
-					.where("key", "=", "non_existent"),
-			),
-		{ wrapper },
-	);
+	let hookResult: { current: State<KeyValue> | undefined };
 
-	await waitFor(() => {
-		expect(result.current.loading).toBe(false);
+	await act(async () => {
+		const { result } = renderHook(
+			() => {
+				const data = useQueryTakeFirst((lix) =>
+					lix.db
+						.selectFrom("key_value")
+						.selectAll()
+						.where("key", "=", "non_existent"),
+				);
+				return data;
+			},
+			{ wrapper },
+		);
+		hookResult = result;
 	});
 
-	expect(result.current.data).toBeUndefined();
+	await waitFor(() => {
+		expect(hookResult.current).toBeUndefined();
+	});
 
 	await lix.close();
 });
 
-test("useQuery vs useQueryTakeFirst loading behavior with empty results", async () => {
+test("useSuspenseQuery return type is properly typed", async () => {
 	const lix = await openLix({});
 	const wrapper = ({ children }: { children: React.ReactNode }) => (
-		<LixProvider lix={lix}>{children}</LixProvider>
+		<LixProvider lix={lix}>
+			<Suspense fallback={<div>Loading...</div>}>
+				<MockErrorBoundary>{children}</MockErrorBoundary>
+			</Suspense>
+		</LixProvider>
 	);
 
-	// Test useQuery with empty results - should return [] and loading=false
-	const { result: queryResult } = renderHook(
-		() =>
-			useQuery((lix) =>
-				lix.db
-					.selectFrom("key_value")
-					.selectAll()
-					.where("key", "=", "empty_test_query"),
-			),
-		{ wrapper },
-	);
+	let hookResult: { current: State<KeyValue>[] } | undefined;
 
-	// Test useQueryTakeFirst with empty results - should return undefined and loading=false
-	const { result: takeFirstResult } = renderHook(
-		() =>
-			useQueryTakeFirst((lix) =>
-				lix.db
-					.selectFrom("key_value")
-					.selectAll()
-					.where("key", "=", "empty_test_takefirst"),
-			),
-		{ wrapper },
-	);
-
-	// Wait for both to complete
-	await waitFor(() => {
-		expect(queryResult.current.loading).toBe(false);
+	await act(async () => {
+		const { result } = renderHook(
+			() => {
+				const data = useQuery((lix) =>
+					lix.db.selectFrom("key_value").selectAll(),
+				);
+				return data;
+			},
+			{ wrapper },
+		);
+		hookResult = result;
 	});
 
+	// Wait for data to be available
 	await waitFor(() => {
-		expect(takeFirstResult.current.loading).toBe(false);
+		expect(hookResult).toBeDefined();
+		expect(Array.isArray(hookResult!.current)).toBe(true);
 	});
 
-	// Verify useQuery returns empty array
-	expect(queryResult.current.data).toEqual([]);
-	expect(queryResult.current.error).toBe(null);
-	expect(queryResult.current.loading).toBe(false);
-
-	// Verify useQueryTakeFirst returns undefined
-	expect(takeFirstResult.current.data).toBeUndefined();
-	expect(takeFirstResult.current.error).toBe(null);
-	expect(takeFirstResult.current.loading).toBe(false);
-
-	await lix.close();
-});
-
-test("useQuery return type is properly typed", async () => {
-	const lix = await openLix({});
-	const wrapper = ({ children }: { children: React.ReactNode }) => (
-		<LixProvider lix={lix}>{children}</LixProvider>
-	);
-
-	const { result } = renderHook(
-		() => useQuery((lix) => lix.db.selectFrom("key_value").selectAll()),
-		{ wrapper },
-	);
-
-	// Type test: data should be properly typed as an array of {key: string, value: string}
+	// Type test: data should be properly typed as an array of KeyValue
 	// This should pass without any type errors if the types are working correctly
-	result.current.data satisfies State<KeyValue>[] | undefined;
-
-	// Wait for loading to complete
-	await waitForUpdate(result, (r) => r.loading);
-
-	expect(result.current.loading).toBe(false);
-	expect(result.current.error).toBe(null);
-	expect(Array.isArray(result.current.data)).toBe(true);
+	hookResult!.current satisfies State<KeyValue>[];
 
 	await lix.close();
 });
 
-test("useQuery type guards work with satisfies operator", async () => {
+test("useSuspenseQuery error handling with ErrorBoundary", async () => {
 	const lix = await openLix({});
+	let caught: Error | undefined;
+
+	// Suppress console errors for this test
+	const originalError = console.error;
+	console.error = () => {};
+
 	const wrapper = ({ children }: { children: React.ReactNode }) => (
-		<LixProvider lix={lix}>{children}</LixProvider>
+		<LixProvider lix={lix}>
+			<Suspense fallback={<div>Loading…</div>}>
+				<MockErrorBoundary onError={(e) => (caught = e)}>
+					{children}
+				</MockErrorBoundary>
+			</Suspense>
+		</LixProvider>
 	);
 
-	const { result } = renderHook(
-		() => useQuery((lix) => lix.db.selectFrom("key_value").selectAll()),
-		{ wrapper },
-	);
-
-	// Wait for loading to complete
-	await waitFor(() => {
-		expect(result.current.loading).toBe(false);
+	await act(async () => {
+		renderHook(
+			() =>
+				useQuery((lix) =>
+					// invalid table: will reject then throw
+					lix.db.selectFrom("non_existent_table" as never).selectAll(),
+				),
+			{ wrapper },
+		);
 	});
 
-	// Type guard tests with satisfies
-	if (result.current.error) {
-		// In error state
-		result.current.data satisfies undefined;
-		result.current.error satisfies Error;
-		result.current.loading satisfies false;
-	} else if (result.current.loading) {
-		// In loading state
-		result.current.data satisfies undefined;
-		result.current.error satisfies null;
-		result.current.loading satisfies true;
-	} else {
-		// Success state - data should be defined array
-		result.current.data satisfies State<KeyValue>[];
-		result.current.error satisfies null;
-		result.current.loading satisfies false;
+	await waitFor(() => {
+		expect(caught).toBeDefined();
+	});
 
-		// TypeScript should know data is not undefined here
-		const length = result.current.data.length; // No ! needed
-		expect(typeof length).toBe("number");
-	}
+	expect(caught!.message).toMatch(/no such table/i);
+
+	// Restore console.error
+	console.error = originalError;
+	await lix.close();
+});
+
+test("useSuspenseQueryTakeFirstOrThrow returns data when result exists", async () => {
+	const lix = await openLix({});
+
+	// Insert test data
+	await lix.db
+		.insertInto("key_value")
+		.values({ key: "throw_test", value: "exists" })
+		.execute();
+
+	const wrapper = ({ children }: { children: React.ReactNode }) => (
+		<LixProvider lix={lix}>
+			<Suspense fallback={<div>Loading...</div>}>
+				<MockErrorBoundary>{children}</MockErrorBoundary>
+			</Suspense>
+		</LixProvider>
+	);
+
+	let hookResult: { current: State<KeyValue> };
+
+	await act(async () => {
+		const { result } = renderHook(
+			() => {
+				const data = useQueryTakeFirstOrThrow((lix) =>
+					lix.db
+						.selectFrom("key_value")
+						.selectAll()
+						.where("key", "=", "throw_test"),
+				);
+				return data;
+			},
+			{ wrapper },
+		);
+		hookResult = result;
+	});
+
+	await waitFor(() => {
+		expect(hookResult.current).toMatchObject({
+			key: "throw_test",
+			value: "exists",
+		});
+	});
 
 	await lix.close();
 });
 
-test("useQueryTakeFirst type guards work with satisfies operator", async () => {
+test("useSuspenseQueryTakeFirstOrThrow throws when no result found", async () => {
 	const lix = await openLix({});
+	let caught: Error | undefined;
+
+	// Suppress console errors for this test
+	const originalError = console.error;
+	console.error = () => {};
+
 	const wrapper = ({ children }: { children: React.ReactNode }) => (
-		<LixProvider lix={lix}>{children}</LixProvider>
+		<LixProvider lix={lix}>
+			<Suspense fallback={<div>Loading...</div>}>
+				<MockErrorBoundary onError={(e) => (caught = e)}>
+					{children}
+				</MockErrorBoundary>
+			</Suspense>
+		</LixProvider>
 	);
 
-	const { result } = renderHook(
-		() =>
-			useQueryTakeFirst((lix) =>
-				lix.db
-					.selectFrom("key_value")
-					.selectAll()
-					.where("key", "=", "nonexistent"),
-			),
-		{ wrapper },
-	);
-
-	// Wait for loading to complete
-	await waitFor(() => {
-		expect(result.current.loading).toBe(false);
+	await act(async () => {
+		renderHook(
+			() => {
+				const data = useQueryTakeFirstOrThrow((lix) =>
+					lix.db
+						.selectFrom("key_value")
+						.selectAll()
+						.where("key", "=", "does_not_exist"),
+				);
+				return data;
+			},
+			{ wrapper },
+		);
 	});
 
-	// Type guard tests with satisfies
-	if (result.current.error) {
-		// In error state
-		result.current.data satisfies undefined;
-		result.current.error satisfies Error;
-		result.current.loading satisfies false;
-	} else if (result.current.loading) {
-		// In loading state
-		result.current.data satisfies undefined;
-		result.current.error satisfies null;
-		result.current.loading satisfies true;
-	} else {
-		// Success state - data can be TResult | undefined (for empty results)
-		result.current.data satisfies State<KeyValue> | undefined;
-		result.current.error satisfies null;
-		result.current.loading satisfies false;
+	await waitFor(() => {
+		expect(caught).toBeDefined();
+		expect(caught!.message).toBe("No result found");
+	});
 
-		// TypeScript should know error is null and loading is false
-		expect(result.current.error).toBe(null);
-		expect(result.current.loading).toBe(false);
-		// In this test case, data should be undefined since we're querying nonexistent key
-		expect(result.current.data).toBeUndefined();
-	}
+	// Restore console.error
+	console.error = originalError;
+	await lix.close();
+});
+
+test("useSuspenseQuery re-executes when query function changes (dependency array fix)", async () => {
+	const lix = await openLix({});
+
+	// Insert test data with different prefixes
+	await lix.db
+		.insertInto("key_value")
+		.values([
+			{ key: "prefix_a_1", value: "value_a_1" },
+			{ key: "prefix_a_2", value: "value_a_2" },
+			{ key: "prefix_b_1", value: "value_b_1" },
+			{ key: "prefix_b_2", value: "value_b_2" },
+		])
+		.execute();
+
+	const wrapper = ({ children }: { children: React.ReactNode }) => (
+		<LixProvider lix={lix}>
+			<Suspense fallback={<div>Loading...</div>}>
+				<MockErrorBoundary>{children}</MockErrorBoundary>
+			</Suspense>
+		</LixProvider>
+	);
+
+	// State to control which prefix to query
+	let hookResult: { current: State<KeyValue>[] };
+	let rerender: (props?: { prefix: string }) => void;
+
+	await act(async () => {
+		const { result, rerender: rerenderFn } = renderHook(
+			({ prefix = "prefix_a" }: { prefix?: string } = {}) => {
+				// Create a new query function each time prefix changes
+				const data = useQuery((lix) =>
+					lix.db
+						.selectFrom("key_value")
+						.selectAll()
+						.where("key", "like", `${prefix}_%`)
+						.orderBy("key", "asc"),
+				);
+				return data;
+			},
+			{
+				wrapper,
+				initialProps: { prefix: "prefix_a" },
+			},
+		);
+		hookResult = result;
+		rerender = rerenderFn;
+	});
+
+	// Wait for initial data (prefix_a results)
+	await waitFor(() => {
+		expect(hookResult.current).toHaveLength(2);
+		expect(hookResult.current[0]).toMatchObject({
+			key: "prefix_a_1",
+			value: "value_a_1",
+		});
+		expect(hookResult.current[1]).toMatchObject({
+			key: "prefix_a_2",
+			value: "value_a_2",
+		});
+	});
+
+	// Change the query prefix - this should trigger a re-execution
+	await act(async () => {
+		rerender({ prefix: "prefix_b" });
+	});
+
+	// Wait for the query to re-execute with new prefix
+	await waitFor(() => {
+		expect(hookResult.current).toHaveLength(2);
+		expect(hookResult.current[0]).toMatchObject({
+			key: "prefix_b_1",
+			value: "value_b_1",
+		});
+		expect(hookResult.current[1]).toMatchObject({
+			key: "prefix_b_2",
+			value: "value_b_2",
+		});
+	});
+
+	await lix.close();
+});
+
+test("useSuspenseQuery subscription updates when query dependencies change", async () => {
+	const lix = await openLix({});
+
+	// Insert initial test data
+	await lix.db
+		.insertInto("key_value")
+		.values([
+			{ key: "sub_test_a_1", value: "initial_a" },
+			{ key: "sub_test_b_1", value: "initial_b" },
+		])
+		.execute();
+
+	const wrapper = ({ children }: { children: React.ReactNode }) => (
+		<LixProvider lix={lix}>
+			<Suspense fallback={<div>Loading...</div>}>
+				<MockErrorBoundary>{children}</MockErrorBoundary>
+			</Suspense>
+		</LixProvider>
+	);
+
+	let hookResult: { current: State<KeyValue>[] };
+	let rerender: (props?: { filter: string }) => void;
+
+	await act(async () => {
+		const { result, rerender: rerenderFn } = renderHook(
+			({ filter = "sub_test_a" }: { filter?: string } = {}) => {
+				const data = useQuery((lix) =>
+					lix.db
+						.selectFrom("key_value")
+						.selectAll()
+						.where("key", "like", `${filter}_%`),
+				);
+				return data;
+			},
+			{
+				wrapper,
+				initialProps: { filter: "sub_test_a" },
+			},
+		);
+		hookResult = result;
+		rerender = rerenderFn;
+	});
+
+	// Verify initial subscription works
+	await waitFor(() => {
+		expect(hookResult.current).toHaveLength(1);
+		expect(hookResult.current[0]?.key).toBe("sub_test_a_1");
+	});
+
+	// Switch to different filter - new subscription should be created
+	await act(async () => {
+		rerender({ filter: "sub_test_b" });
+	});
+
+	// Verify new subscription works
+	await waitFor(() => {
+		expect(hookResult.current).toHaveLength(1);
+		expect(hookResult.current[0]?.key).toBe("sub_test_b_1");
+	});
+
+	// Insert new data that matches the current filter
+	await act(async () => {
+		await lix.db
+			.insertInto("key_value")
+			.values({ key: "sub_test_b_2", value: "new_b" })
+			.execute();
+	});
+
+	// The subscription should pick up the new data
+	await waitFor(() => {
+		expect(hookResult.current).toHaveLength(2);
+		expect(hookResult.current.some((item) => item.key === "sub_test_b_2")).toBe(
+			true,
+		);
+	});
 
 	await lix.close();
 });
