+++ packages/lix/lix-react-utils/src/index.ts
@@ -1,8 +1,10 @@
 export { LixProvider, LixContext } from "./provider.js";
-export { useQuery, useQueryTakeFirst } from "./hooks/use-query.js";
 export {
+	useQuery,
+	useQueryTakeFirst,
+	useQueryTakeFirstOrThrow,
 	useSuspenseQuery,
 	useSuspenseQueryTakeFirst,
 	useSuspenseQueryTakeFirstOrThrow,
-} from "./hooks/use-suspense-query.js";
+} from "./hooks/use-query.js";
 export { useLix } from "./hooks/use-lix.js";
